Namespace the RakeMappings in the same vein as RubyMappings
